courseapp/requirement.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.38.0
colorama==0.4.6
cryptography==41.0.7
decorator==5.1.1
Deprecated==1.2.14
Django==5.0
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
executing==2.0.1
idna==3.6
inflection==0.5.1
ipython==8.20.0
jedi==0.19.1
jwcrypto==1.5.0
libopencv==0.0.1
matplotlib-inline==0.1.6
mysqlclient==2.2.1
numpy==1.23.5
oauth2-provider==0.0
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==23.2
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycparser==2.21
pygments==2.17.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.13
wrapt==1.16.0
